More packet switchovers, NpcInOutForMe fix and looting fixed
Switched over some more packets, fixed a problem where a required part
of the NpcInOutForMe packet wasn't constructed properly because that
particular line ended up in a comment, fixed looting. Rather odd thing
is that even though the server doesn't have items like mana potions in
the item_bundle that's being opened (and doesn't send it to the client!)
they still show up in loot boxes - not quite sure what to do with this.